Popup Relevance fix
expnasion panel dialog fix
