+++ android-json-form-wizard/src/main/java/com/vijay/jsonwizard/fragments/JsonWizardFormFragment.java
@@ -155,6 +155,12 @@ public class JsonWizardFormFragment extends JsonFormFragment {
         }
 
         setJsonFormFragment(this);
+
+        try {
+            Thread.sleep(100);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
     }
 
     @Override
