+++ gradle.properties
@@ -1,4 +1,4 @@
-VERSION_NAME=1.9.2-r36-OPTIMIZED-LOCAL-SNAPSHOT
+VERSION_NAME=1.9.2-r84-OPTIMIZED-LOCAL-SNAPSHOT
 VERSION_CODE=1
 GROUP=org.smartregister
 POM_SETTING_DESCRIPTION=OpenSRP Client Native Form Json Wizard
